Finally fixed the app so its working and no 10000 errors but the new function still alot problem

+++ pawsure_app/lib/screens/calendar/calendar_screen.dart
@@ -6,7 +6,6 @@ import 'package:pawsure_app/controllers/calendar_controller.dart';
 import 'package:pawsure_app/controllers/home_controller.dart';
 import 'package:pawsure_app/models/event_model.dart';
 import 'package:pawsure_app/widgets/calendar/add_event_modal.dart';
-import 'package:pawsure_app/widgets/calendar/edit_event_modal.dart';
 
 class CalendarScreen extends StatelessWidget {
   const CalendarScreen({super.key});
@@ -37,7 +36,7 @@ class CalendarScreen extends StatelessWidget {
                     (p) => p.id == petId,
                   );
                   homeController.selectPet(pet);
-                  controller.loadEvents(petId as int);
+                  controller.loadEvents(petId);
                 },
                 itemBuilder: (context) => homeController.pets
                     .map(
